refactor(hooks/builtins): inline readPromptFiles into add_prompt_files
The session.ReadPromptFiles helper had only one caller — the
add_prompt_files builtin — so move it (and its private
findFileInHierarchy / isFile helpers) into the builtin itself as
unexported functions and drop the cross-package dependency.

- pkg/hooks/builtins/add_prompt_files.go: inlines readPromptFiles,
  findFileInHierarchy, isFile; removes the pkg/session import.
- pkg/hooks/builtins/add_prompt_files_test.go: relocates the seven
  prompt-file behavior tests as internal tests so they can exercise
  the now-unexported readPromptFiles directly.
- pkg/session/prompt_file.go and prompt_file_test.go: removed.

No behavior change: the public add_prompt_files hook keeps the same
working-dir-hierarchy + home-folder lookup, dedup, and missing-file
tolerance.
